+++ src/main/java/frc/robot/constants/Constants.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.constants;
 
 public class Constants {
 	public static final double liftEncoderPerInch = 914;
