+++ src/main/java/frc/robot/constants/ElectricalConstants.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.constants;
 
 public class ElectricalConstants {
 	// Controllers
