+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -1,7 +1,8 @@
-package frc.robot;
+package frc.robot.subsystems;
 
-// import edu.wpi.first.wpilibj.DigitalInput;
-// import com.ctre.phoenix.motorcontrol.can.*;
+import frc.robot.constants.ElectricalConstants;
+import frc.robot.drivebase.Controls;
+import frc.robot.constants.Constants;
 
 public class Intake{
